RUI-4 feat: Input 컴포넌트 수정 - 타입 분리 - Hint Text Color를 임의로 지정 가능 - left, right icon 추가 방식 변경
